deployment/requirements.txt
flask
langchain_openai
langgraph
requests
python-dotenv 